+++ ui/base/ime/linux/input_method_auralinux_unittest.cc
@@ -239,7 +239,7 @@ class TextInputClientForTesting : public DummyTextInputClient {
 
   bool HasCompositionText() const override { return !composition_text.empty(); }
 
-  uint32_t ConfirmCompositionText(bool keep_selection) override {
+  size_t ConfirmCompositionText(bool keep_selection) override {
     // TODO(b/134473433) Modify this function so that when keep_selection is
     // true, the selection is not changed when text committed
     if (keep_selection) {
@@ -247,8 +247,7 @@ class TextInputClientForTesting : public DummyTextInputClient {
     }
     TestResult::GetInstance()->RecordAction(u"compositionend");
     TestResult::GetInstance()->RecordAction(u"textinput:" + composition_text);
-    const uint32_t composition_text_length =
-        static_cast<uint32_t>(composition_text.length());
+    const size_t composition_text_length = composition_text.length();
     composition_text.clear();
     return composition_text_length;
   }
